// HRMS-MATERIAL-UI/src/app/modules/reports/models/projects-report.model.ts
export class ProjectResponse{
    ProjectId: number;
    ProjectName: string;
    PlannedStartDate: Date;
    PlannedEndDate: Date;
    ActualStartDate: string;
    ActualEndDate: string;
    ClientId: number;
    ClientName: string;
    PracticeAreaCode: string;
    ProjectTypeCode: string;
    ProjectTypeDescription: string;
    StatusCode: string;
    StatusId: number;
    ProjectStateId: number;
    ProjectState: string;
    ManagerName: string;
    ManagerId: number;
    PracticeAreaId: number;
    ProjectTypeId: number;
    DepartmentId: number;
    DepartmentCode: string;
    DomainId: number;
    DomainName: string;
    IsActive: boolean
    ProgramManagerId: number;
    UserRole: string;
    ProgramManager: string;
    ReportingManagerId: number;
    EmployeeCode: string;
    EmployeeId: number;
    FirstName: string;
    LastName: string;
    PracticeArea: string;
}

// "ProjectId": 89,
//     "ProjectCode": "P000002009",
//     "ProjectName": "Access Card Maintenance Tool",
//     "PlannedStartDate": null,
//     "PlannedEndDate": null,
//     "ActualStartDate": "2020-02-01T00:00:00+05:30",
//     "ActualEndDate": null,
//     "ClientId": 21,
//     "ClientName": "SenecaGlobal",
//     "PracticeAreaCode": "Development (EM)",
//     "ProjectTypeCode": null,
//     "ProjectTypeDescription": null,
//     "StatusCode": "Execution",
//     "StatusId": 19,
//     "ProjectStateId": 19,
//     "ProjectState": "Execution",
//     "ManagerName": "Sasfname LnameDat",
//     "ManagerId": 152,
//     "PracticeAreaId": 5,
//     "ProjectTypeId": 0,
//     "DepartmentId": 1,
//     "DepartmentCode": "Delivery",
//     "DomainId": 0,
//     "DomainName": null,
//     "IsActive": null,
//     "ProgramManagerId": 0,
//     "UserRole": null,
//     "ProgramManager": null,
//     "ReportingManagerId": null,
//     "EmployeeCode": null,
//     "EmployeeId": 0,
//     "FirstName": null,
//     "LastName": null,
//     "PracticeArea": null