// HRMS-MATERIAL-UI/src/app/modules/reports/models/service-typereport.model.ts
export class ServiceTypeReport {
    ResourceCount: number;
    ServiceTypeId: number;
    ServiceTypeDescription: string;
    ServiceTypeName: string;
}

export class ServiceTypeReportEmp {
    EmployeeCode: string;
    EmployeeId: number;
    EmployeeName: string;
    Experience: number;
    Designation: string;
    Grade: string;
    ProjectName: string;
}

export class ServiceTypeDropDown {
    Id: number;
    Name: string;
    Description: string;
    Code: string;
}



  